Create oneful pairs
Chef defines a pair of positive integers 
(
𝑎
,
𝑏
)
(a,b) to be a 
Oneful Pair
Oneful Pair, if

𝑎
+
𝑏
+
(
𝑎
⋅
𝑏
)
=
111
a+b+(a⋅b)=111

For example, 
(
1
,
55
)
(1,55) is a 
Oneful Pair
Oneful Pair, since 
1
+
55
+
(
1
⋅
55
)
=
56
+
55
=
111
1+55+(1⋅55)=56+55=111.
But 
(
1
,
56
)
(1,56) is not a 
Oneful Pair
Oneful Pair, since 
1
+
56
+
(
1
⋅
56
)
=
57
+
56
=
113
≠
111
1+56+(1⋅56)=57+56=113=111.

Given two positive integers 
𝑎
a and 
𝑏
b, output Yes if they are a 
Oneful Pair
Oneful Pair. And No otherwise.